requirements.txt
# ==================================================================================
# VAUCDA Backend Requirements
# Python 3.11+
# ==================================================================================

# Web Framework & ASGI Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
websockets==12.0

# Async Support
aiofiles==23.2.1
aiosqlite==0.19.0
httpx==0.25.2

# Data Validation & Settings
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Database Drivers
neo4j==5.15.0
neo4j-graphdatabase==5.15.0

# LLM Integration
ollama-python==0.1.0
anthropic==0.18.0
openai==1.10.0

# LangChain & RAG
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2
langchain-anthropic==0.0.1
langchain-neo4j==0.1.0

# Embeddings & Vector Search
sentence-transformers==2.2.2
torch==2.1.2
transformers==4.36.2

# Document Processing
pymupdf==1.23.8
python-docx==1.1.0
unstructured==0.11.0
pypdf==3.17.4

# Task Queue
celery[redis]==5.3.4
redis==5.0.1
kombu==5.3.4

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7
python-multipart==0.0.6

# Utilities
python-dateutil==2.8.2
pytz==2023.3
pendulum==3.0.0

# HTTP & API
requests==2.31.0
httpx==0.25.2

# Logging & Monitoring
python-json-logger==2.0.7
prometheus-client==0.19.0

# Data Processing
numpy==1.26.2
pandas==2.1.4
scipy==1.11.4

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
faker==21.0.0

# Code Quality
black==23.12.1
pylint==3.0.3
mypy==1.7.1
isort==5.13.2

# Development
ipython==8.19.0
ipdb==0.13.13
